Replace reqwest with much leaner minreq
It is possible to use minreq 1.0.0, which would make the minimum rustc
version 1.30. However, there seems to be a bug in this version of minreq
which occasionally causes tests to fail. It appears the problem is gone
in minreq 1.1, which however requires at least rustc 1.31 to compile, so
that is the minimum rustc version required from now one.