Split library and tests to separate crates
This commit prepares for making various dependencies optional. Due to limitation
in cargo, it is not possible to enable specific features for tests. As a
workaround, integration tests are in separate crate that uses vk-parse with
required features enabled. Everything is put together using virtual manifest in
the root of the repository. Finally, .travis.yml is modified to use custom
script for testing whole workspace.