Untrack "unused" Cargo.lock file
As discussed in [1] tracking `Cargo.lock` does not help dependent crates
adhere to `minimal-versions`, that only applies to the CI here.  As such
it makes more sense to drop `Cargo.lock` and "generate" it on the fly in
the CI to make sure that `Cargo.toml` - which _is_ read by dependent
crates - adheres to `minimal-versions` at all times.

[1]: https://github.com/krolli/vk-parse/pull/20