parse_interface_item: fix passing over unrecognized elements
previously, if things that utilized this separated match pattern would
fail if encountering any unexpected/unknown elements.

EX. when parsing the OpenXR spec, `require` can contain
`interaction_profile` elements, which causes `parse_interface_item` to
return None (as an error case), and the rest of the block is never
processed.

I introduced an alternative pattern to `unwrap_attribute!` in
`parse_attributes!` that allows to acheive a simliar pattern, without
the early return being *in the macro*, allowing the invoker to regain
control of the control flow.

This allowed moving all invocations of `parse_interface_item` to more
standard `parse_*` function patterns used elsewhere, allowing it to
function simliarly and properly handle unexpected elements.